Split the Solver header into actual solver + config
